Refactor light sleep handling and improve MQTT configuration management
Update the architecture and module documentation to clarify the changes in light sleep wakeup sources, separating static and timer-based wakeup functions. Adjust MQTT port type to `uint16_t` for better range validation and enhance the configuration loading logic. Additionally, refine display initialization based on debug level and optimize heart counter management in the MQTT callback for improved performance and reliability.

+++ src/config.h
@@ -1,7 +1,9 @@
 #pragma once
 
+#include <cstdint>
+
 extern char mqtt_server[128];
-extern int mqtt_port;
+extern uint16_t mqtt_port;
 extern char mqtt_username[64];
 extern char mqtt_password[64];
 extern char mqtt_topic_pub[128];
